Asyncify connection handling
The program still works as it used to, but now it's ready to be
converted over to an event-loop / async style. Who needs async/await
when you've got manually managed structs that do the same thing ;)

Real talk: wouldn't recommend manual async, it was a big pain to get
working, and it's probably riddled with bugs. Just use Rust for
servers.